middle-end/113680 - Optimize (x - y) CMP 0 as x CMP y
(x - y) CMP 0 is equivalent to x CMP y where x and y are signed
integers and CMP is <, <=, >, or >=.  Similarly, 0 CMP (x - y) is
equivalent to y CMP x.  As reported in PR middle-end/113680, this
equivalence does not hold for types other than signed integers.  When
it comes to conditions, the former was translated to a combination of
sub and test, whereas the latter was translated to a single cmp.
Thus, this optimization pass tries to optimize the former to the
latter.

When `-fwrapv` is enabled, GCC treats the overflow of signed integers
as defined behavior, specifically, wrapping around according to two's
complement arithmetic.  This has implications for optimizations that
rely on the standard behavior of signed integers, where overflow is
undefined.  Consider the example given:

	long long llmax = __LONG_LONG_MAX__;
	long long llmin = -llmax - 1;

Here, `llmax - llmin` effectively becomes `llmax - (-llmax - 1)`, which
simplifies to `2 * llmax + 1`.  Given that `llmax` is the maximum value
for a `long long`, this calculation overflows in a defined manner
(wrapping around), which under `-fwrapv` is a legal operation that
produces a negative value due to two's complement wraparound.
Therefore, `llmax - llmin < 0` is true.

However, the direct comparison `llmax < llmin` is false since `llmax`
is the maximum possible value and `llmin` is the minimum.  Hence,
optimizations that rely on the equivalence of `(x - y) CMP 0` to
`x CMP y` (and vice versa) cannot be safely applied when `-fwrapv` is
enabled.  This is why this optimization pass is disabled under
`-fwrapv`.

This optimization pass must run before the Jump Threading pass and the
VRP pass, as it may modify conditions. For example, in the VRP pass:

	(1)
	  int diff = x - y;
	  if (diff > 0)
	    foo();
	  if (diff < 0)
	    bar();

The second condition would be converted to diff != 0 in the VRP pass
because we know the postcondition of the first condition is diff <= 0,
and then diff != 0 is cheaper than diff < 0. If we apply this pass
after this VRP, we get:

	(2)
	  int diff = x - y;
	  if (x > y)
	    foo();
	  if (diff != 0)
	    bar();

This generates sub and test for the second condition and cmp for the
first condition. However, if we apply this pass beforehand, we simply
get:

	(3)
	  int diff = x - y;
	  if (x > y)
	    foo();
	  if (x < y)
	    bar();

In this code, diff will be eliminated as a dead code, and sub and test
will not be generated, which is more efficient.

For the Jump Threading pass, without this optimization pass, (1) and
(3) above are recognized as different, which prevents TCO.

	PR middle-end/113680

gcc/ChangeLog:

	* Makefile.in: Add tree-ssa-cmp.o to OBJS.
	* common.opt: Define ftree-cmp
	* doc/invoke.texi: Document ftree-cmp.
	* opts.cc (default_options_table): Handle OPT_ftree_cmp.
	* passes.def (pass_cmp): New optimization pass.
	* timevar.def (TV_TREE_CMP): New variable for timing.
	* tree-pass.h (make_pass_cmp): New declaration.
	* tree-ssa-cmp.cc: New file.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-ssa/pr113680.c: New test.

Signed-off-by: Ken Matsui <kmatsui@gcc.gnu.org>